Cleanup - reduced code duplication
